add title

+++ main.py
@@ -1,6 +1,5 @@
 import aiogram
 import os
-import aiogram.filters
 import dotenv
 import asyncio
 import yt_dlp
@@ -12,26 +11,40 @@ bot = aiogram.Bot(os.getenv("BOT_TOKEN"))
 dp = aiogram.Dispatcher()
 
 
+def is_valid_url(url: str):
+    if url.startswith("https://youtube.com/"):
+        return True
+    else:
+        return False
+
+
 @dp.message()
 async def start_handler(message: aiogram.types.Message):
-   url = message.text 
-   filename = random.randint(1,100)
-   opts = {"format" : "best" , "outtmpl" : f"{filename}.%(ext)s"}
+    url = message.text
 
-   yt_dlp.YoutubeDL(opts).download(url)
+    if is_valid_url(url):
+        filename = random.randint(1, 100)
+        opts = {"format": "best", "outtmpl": f"{filename}.%(ext)s"}
 
-   video = aiogram.types.FSInputFile(f"{filename}.mp4")
+        await message.answer("video is downloading, please wait a moment...")
 
-   await message.answer_video(video)
+        video_info = yt_dlp.YoutubeDL(opts).extract_info(url, download=True)
 
-   os.remove(f"{filename}.mp4")
+        video = aiogram.types.FSInputFile(f"{filename}.mp4")
+        await message.answer_video(video, caption=video_info["title"])
+
+        os.remove(f"{filename}.mp4")
+    else:
+        await message.answer(f"{url} is not valid url.")
+
 
 def on_start():
     print("bot has been started...")
+
 
 async def main():
     dp.startup.register(on_start)
     await dp.start_polling(bot)
 
-asyncio.run(main())
 
+asyncio.run(main())
